fix_redirection_cat_commands
