+++ srcs/executor/get_path.c
@@ -36,32 +36,10 @@ int	init(char ***path_splitted)
 	return (EXIT_SUCCESS);
 }
 
-static char	*get_abs_cmd_path(char **abs_cmd_path,
-		char *path_splitted, char *cmd)
-{
-	*abs_cmd_path = ft_strjoin2(path_splitted, "/");
-	if (abs_cmd_path == NULL)
-		return (NULL);
-	*abs_cmd_path = ft_strjoin(*abs_cmd_path, cmd);
-	if (*abs_cmd_path == NULL)
-		return (*abs_cmd_path);
-	return (*abs_cmd_path);
-}
-
 void	free_path(char **dup, char **abs_path)
 {
 	free_me(dup);
 	free_me(abs_path);
-}
-
-char	*get_2d_path(char *abs, char **path_split, int i, char *cmd)
-{
-	char	**tmp;
-
-	tmp = ft_split(cmd, ' ');
-	abs = get_abs_cmd_path(&abs, path_split[i], tmp[0]);
-	free_2d_array(tmp);
-	return (abs);
 }
 
 char	*get_abs_cmd(char *cmd)
